refactor(llm): hoist LLMLayer + ax_runner_t into LLMLayer.hpp (KvSlotManager stage 3a)
Move the LLMLayer struct and the compile-time ax_runner_t backend typedef (+ its
backend-runner includes) out of LLM.cpp into src/runner/LLMLayer.hpp. Pure move;
no behavior change. Prerequisite for stage 3b: KvSlotManager will hold a
std::vector<LLMLayer>& and so needs LLMLayer visible in a header.

Verified: LLM.cpp compiles with real build_x86 (AXCL) flags on the bridge
(EXIT 0); braces balanced.
